Scikit-learn SVM parameters and handwritten digits
***Handwritten Digits***
* This dataset is a dataset that ships with scikit-learn. The
handwritten digits datasets was transformed into
"8x8 images of integer pixels in the range 0..16." More about it
described here:
http://scikit-learn.org/stable/datasets/index.html#optical-recognition-of-handwritten-digits-data-set
* Since this data is already normalized, we do not need to perform
feature scaling or anything like that.

***SVM***
* Support Vector Machines
* Default kernel="rbf" = radial basis function = Gaussian kernel

***Shuffling***
* We are dividing the data into a few groups so we would like to ensure
the order is random so that our groups are not bias

***Accuracy Score***
* Percentage of correct predictions. This assesses how well the model
predicts data in the test or cross validation set. However, this scoring
becomes invalid if the data set were to be skewed towards a certain
classification. For example, if most of the digits were 2s.

***Training, Cross Validation, Test***
* The training set is used to actually train the model
* The cross validation set is used to tweak parameters (in the case of
Gaussian kernel SVM in scikit learn we tweak gamma and C)
* The test set tells us how accurate our model is