Pick the best regularization parameter, alpha
* The MLPClassifier takes an optional regularization parameter. This
is parameter is used to prevent "overfitting" the training data. That
means the predictions perform well on the training set, but fail to
generalize to the cross validation (or test set).
* Here we implement another `h.accuracy_scores_for_nn` function to try a
range of alpha values.
* In the graph we can see the effects of underfitting (setting alpha to
high) but the effects of overfitting is not as bad here. The best value
for alpha seems to continually be "1", and I am uncertain as to why that
is.